+++ lib/src/flutter_shield_platform_interface.dart
@@ -61,6 +61,9 @@ abstract class FlutterShieldPlatform extends PlatformInterface {
   Future<SecurityCheckResult> checkDeviceTime();
   Future<SecurityCheckResult> checkSideChannel();
 
+  // Play Integrity (Android only)
+  Future<SecurityCheckResult> checkPlayIntegrity();
+
   // Comprehensive check
   Future<SecurityReport> performFullSecurityCheck();
 }