+++ lib/src/flutter_shield_stub.dart
@@ -138,6 +138,10 @@ class FlutterShieldStub extends FlutterShieldPlatform {
       _notApplicable(VulnerabilityType.sideChannelAttacks);
 
   @override
+  Future<SecurityCheckResult> checkPlayIntegrity() async =>
+      _notApplicable(VulnerabilityType.playIntegrityFailed);
+
+  @override
   Future<SecurityReport> performFullSecurityCheck() async {
     final results = await Future.wait([
       checkRootedJailbroken(),
@@ -171,6 +175,7 @@ class FlutterShieldStub extends FlutterShieldPlatform {
       checkSensorAbuse(),
       checkDeviceTime(),
       checkSideChannel(),
+      checkPlayIntegrity(),
     ]);
 
     return SecurityReport(